device_mesh: move to own file + loosen input types
